WIP: Fix create account page progress...
Tried to make a function to check the data and submit. An error is happening where the code I used the get the form data for the login screen isn't working, I think I need to debug this however I can't do that on my work laptop. Hopefully once I've fixed the bug it will work the way it's supposed to.